pyproject.toml
[tool.poetry]
name = "birdtrack"
version = "0.1.0"
description = ""
authors = ["inogeo <pierre.hardy5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.7"
pyyaml = "^6.0.2"

[poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
